Don't require ancestor chains in validators. (#4153)
## Motivation

Currently a validator cannot process a child chain without syncing all
the ancestor chains up to the point where the child chain was created.

## Proposal

When a client uploads the first block of a chain, the validator should
look up the committee and verify the signatures. It should then, based
on that, accept the chain description blob, even if it doesn't have the
parent chain.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- closes #4085