Implement the accounting of runtime fees. (#4125)
## Motivation

The access to runtime entries like `ChainId` is not expensive, but it
has to be accounted for.

Fixes #2905 

## Proposal

The costs imposed on the users have to be proportional to what it costs
the validators. Also,
they have to be easy to understand by the developer.

The following design is then done:
* The user is charged a cheap cost proportional to the number of bytes
that they obtain.
* For `ChainId`, `ApplicationId`, and `Timestamp`, that is easy.
However, for a structure like the `ChainOwnership`, this is more
complicated since the size depends on the implementation. So, instead,
we compute a size that is easy to understand by the developer.
* For the parameters, this is easy to do.

Two clone operations are removed in other parts of the code.

## Test Plan

A test has been added for the size of constant-size types.
The CI should handle other costs.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None