Make benchmark client scale better (#4122)
## Motivation

We were needlessly using `spawn_blocking` for a lot of things, with the
`cross_beam` channels... was a bit too complicated.

## Proposal

Simplify the code and make it scale better by using an `AtomicUsize` and
tokio's `Notify`. This should make us be able to use up all the vCPUs
even from a big benchy instance.

## Test Plan

Will deploy a network and benchmark against it using this

## Release Plan

- Nothing to do / These changes follow the usual release cycle.