Make benchmark client design less racey (#4158)
## Motivation

Now that we're gonna start processing all the messages as well, the
current design was very prone to race conditions: blocks for different
chains get executed in parallel, but the blocks actually depend on the
blocks being executed by other chains/tasks, and the messages those
blocks are sending.
So it was very easy for race conditions to happen where a block tries to
execute a message that depends on another message that wasn't received
yet.

## Proposal

Tweak the design a bit:
Instead of having one tokio task per chain, now each tokio task will
instead execute a chain group: several chains, that only send messages
to each other. We always execute one block for one of the chains, then
move on to the next chain in the group, serially. This should prevent
race conditions, as blocks being executed by different tokio tasks don't
depend on each other anymore.

Now you can just specify the bps you want with `--bps` and the size of
your blocks with `--transactions-per-block`, and we'll set the number of
chain groups to be the number of CPUs you have, by default. This seems
to be enough to saturate CPU (at least locally, have to check on GCP) as
you ramp up the BPS amount.

## Test Plan

Deploy a local kind network, run the benchmark against it. I'll also run
this against a GCP network

## Release Plan

- Nothing to do / These changes follow the usual release cycle.