Include pending message bundles in benchmark block proposals (#4150)
## Motivation

We were not processing the inboxes during the benchmarks, which caused
the inboxes to just grow indefinitely.

## Proposal

Include the pending messages with the block proposals. This will cause
each block to do a bit more work, which will likely force us to use
smaller block sizes in the benchmarks to control latency, but will
prevent the server's memory from growing indefinitely.

## Test Plan

Will deploy a network and make sure memory isn't linearly increasing
anymore, as well as track the inbox sizes.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.