Add delay between chain groups for TPS ramp up (#4163)
## Motivation

When running benchmarks, if we're sending a very large TPS amount, it is
useful to be able to ramp up that TPS.

## Proposal

Add an option to add a delay for subsequent chain groups to be started.
This should be set proportionally to the amount of work (TPS) each chain
group will be doing.

## Test Plan

Tested against a local network, saw that the delay works as expected.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.