Eliminate the `NotFound` for the `read_confirmed_block*`. (#4161)
## Motivation

The `ViewError::NotFound` was used in the query. We want to eliminate it
from the `ViewError`. So, this case can be ironed out.

## Proposal

The following is done:
* The `ViewError::not_found` is removed.
* The `read_confirmed_block` returns an option.
* The `read_confirmed_blocks_downward` is eliminated from
`linera-storage` and when needed it is inlined (two contexts, one being
a test).
* The variant is added to the `ChainClientError`. There are no more
functions in `ChainClient` whose error is `ViewError`.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.