Implement `NodeService` balance (#4172)
## Motivation

Like other issues, the end-to-end tests are a great place to test the
fee structure of Linera.
The `ClientWrapper` provides the `fn local_balance` and `fn
query_balance`. However, once the `NodeService``
has been created, they cannot be used. 

## Proposal

The `balance` functionality is added to the `NodeService`:
* It uses GraphQL to access the values.
* For the `Account::chain(...)` the values is accessed via the `balance`
entry. For other the `balances` is used.
* If an account is missing, then the value 0 is returned.

## Test Plan

The CI test of reconfiguration is adapted.
The test of balance is added to the `test_wasm_end_to_end_counter`.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.