Add tokio threads options to shard deployment (#4170)
## Motivation

We'll be starting to set the tokio worker threads in the server binary
from `lineractl`.

## Proposal

Add option for tokio worker threads and blocking threads.

## Test Plan

Deployed a network locally with this, saw that it didn't set the env
variable, as expected.
Deployed a network using `lineractl`, saw that it did set the env
variable, as expected.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.