Implement `msg.sender` in Linera support of the EVM. (#4134)
## Motivation

When calling the EVM smart contract we currently set up the `caller`
entry to `Address::ZERO`.
This is quite inadequate as the `msg.sender` construct is widely used in
solidity code, for example
with ERC20.

Fixes #4129 

## Proposal

The `msg.sender` is first the Externally Owned Account when calling a
contract. When the contract references another contract, the address
becomes that of the EVM contract.

The following is done:
* First of all, we need to set the user's address to an `Address20`
entry. This is done in a somewhat hackish way by modifying the
`generate_from` in `crypto/mod.rs` according to the `revm` feature.
* For `process_streams` there are no `authenticated_*` set so we set it
to `Address::ZERO`.
* Then the `caller` entry is computed from the context in question. That
is the `authenticated_*` entries. For the application id, we require it
to be of one of an EVM address. For a user address, we require it to be
an `Address20`.
* The function `get_msg_address` first tries to access the
application_id, and then if nothing is available, the
`authenticated_caller_if`. This is to implement the `msg.sender`
semantic.
* A wasm contract calling an EVM contract has the address set to
`Address::ZERO`.
* A user in `Address32` calling an EVM smart contract will have
`Address::ZERO`.
* For service calls, the `caller` is set to `SERVICE_ADDRESS`, a
constant address that is distinct from `Address::ZERO`.
* The database is reorganized in order to reflect the fact that now we
have multiple users and multiple nonces to store.

Other notes:
* The next step is to handle the balances so that effective transfers
can be done.
* The solidity smart contracts used for tests are returning a uint64.
This is artificial, as it seems that there is a bug in REVM.


## Test Plan

A test in CI has been added.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.