Leave gaps in sender chains on the validator side (#4181)
## Motivation

If a client creates a block receiving messages from a chain with gaps,
and a validator doesn't have that chain yet, the client currently has no
way of convincing the validator that the proposal is correct, because
the validator expects the whole sender chain as proof of the sent
messages.

## Proposal

Make validators accept confirmed certificates for blocks on chains with
gaps. If there is a gap, blocks should only be preprocessed, instead of
fully processed - like on the clients.

## Test Plan

CI with an added test exercising the problematic scenario

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- closes #4160