Remove the "blst" dependency. (#4197)
## Motivation

The `blst` library is used as a dependency in `revm`, but it has the
disadvantage of being an encapsulation of a "C" library, which prevents
its usage on Wasm.

## Proposal

Remove "blst" as a dependency. This is fine since an examination of the
source code of "revm" reveals that if this feature is not enabled then
"revm" uses the "ark-bls12-381" library. The name of the encapsulation
being `crypto_backend`.

## Test Plan

The CI.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.