Unify confirmed certificate processing between clients and validators (#4195)
## Motivation

Now that validators allow for gaps in chains, we can unify the way
confirmed certificate are processed on the client and on the validator.

## Proposal

Remove the `PreprocessCertificate` worker request and use
`ProcessConfirmedBlock` instead (which can still call
`preprocess_certificate` under the hood if there are gaps).

Moreover, always use the committee from storage for certificate
validation. We used to get the current committee from the chain state
for nonzero-height blocks, but this can be incorrect if there are gaps,
because the committee might have changed within the gap, or we might not
even have a committee in the chain state. Thus, we should always use the
block epoch to get the committee from storage, but if the block is going
to be fully processed, we should also verify that the epoch in it
matches the current epoch on the chain.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)