Add random ID to multi-benchmark in case we want to run several multi-benchmark processes in the same machine (#4177)
## Motivation

Using the same benchy instance with local SSD to benchmark different
networks at the same time is currently not possible because it assumes
one runner per machine (which is the most common use case). However, I
have needed to benchmark different networks at the same time to save
time.

## Proposal

Create a directory with a random name for each instance of the
`multi-benchmark` that is running, within the local ssd directory, to
save the client state (wallet and DB)

## Test Plan

Using this to benchmark several networks at the same time

## Release Plan

- Nothing to do / These changes follow the usual release cycle.