Use NodeConfig to setup Scylla (#4193)
## Motivation

There's this `NodeConfig` that we can use to setup the disks properly
for Scylla, and do a few other optimizations to the setup.

## Proposal

Use the `NodeConfig` for Scylla, but continue using the existing hook
script to setup local SSD for the shards, when using `DualStore`. Make
sure the Scylla stuff only runs in the Scylla VMs by using placements.

## Test Plan

Deployed a network with these changes both using `DualStore` and using
just `ScyllaDB`, both work.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.