Implement ERC20 in a contract that also implement transfers (#4204)
## Motivation

The ERC20 is a major fungible token smart contract in the Ethereum
world.
They form a basis of Decentralized Finance applications such as Morpho,
Uniswap, or others.

Fixes #3931

## Proposal

The reference implementation [OpenZeppelin
ERC20](https://docs.openzeppelin.com/contracts/4.x/erc20) is the one
that is used in most cases. So, we take it as the basis for our
implementation.

In our implementation, we have a constant value. There is no minting or
burning. That implements the `totalSupply` very easily: We simply return
the value at the constructor level.

The transfer from one chain to another is done with one `_burn` (in the
sending chain) and one `_mint` (in the receiving chain).

However, since we have a multi-chain, and we take the total chain, we
have to change the code so that the `totalSupply` would work. One design
that would not have worked is to have a child contract since the events
are emitted with the contract address and the `msg.sender` is used to
identify the sender.
 
The ERC20 contract depends on the `OpenZeppelin`, which is published
via`npm`. Therefore, we have added this to the compilation in
`solidity.rs`

## Test Plan

The CI tests the functionalities.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.