Explicitly set max_concurrent_streams to u32::MAX-1 (#4237)
## Motivation

In the absence of an explicitly set limit, the GCP load balancer was
defaulting to 100, effectively limiting the maximum number of chains in
the wallet to 100 (as the client starts one stream per chain in the
wallet).

## Proposal

Explicitly set the limit to as high a value as possible.

## Test Plan

- CI
- Manually attempt to reproduce #3898 on the devnet after deploying the
fix

## Release Plan

- These changes should be backported to the latest `devnet` branch, then
    - be released in a validator hotfix.
- These changes should be backported to the latest `testnet` branch,
then
    - be released in a validator hotfix.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)