Insert a `DelegatedFungible` smart contract. (#4210)
## Motivation

Both Ethereum with the ERC20 and Solana with the SPL have informal
standards about the minimal set of capabilities of a fungible contract.

In those standards, a user A can allow another user B to access his
funds up to some extent.
Then user B can spend tokens of user A up to the allowance.
This is missing from the fungible app.

## Proposal

Currently, we have `fungible` and `native-fungible` that implement the
same API. A new `delegated-fungible` contract has been introduced. This
is because both for Ethereum and Solana, the fungible token are informal
standard that are not part of the core protocol.

In order to test the smart contract we are forced by the API to have a
multi-owner chain.
This is the first time that a wasm application is deployed on a
multi-owner chain.
This leads to several interesting things:
* The last argument of `publish_and_create` is `Some(chain2)` as opposed
to `None` in all other tests.
* It would be nice to be able to change the default chain in the
`ClientWrapper`.
* The `process_inbox` command is used on the `node_service` even if only
one chain is being used. This is because the first thing that
`process_inbox` does is `synchronize_from_validators`. The
synchronization is available in the client, but not in the
`node_service`.

The `fungible` was taken as a model, and some improvements made in
`delegated-fungible` could be backported to `fungible`:
* The `fn instantiate` does not check for emptiness but for the total
sum of tokens. I think this is better.
* In the case of the absence of tokens, the `fn instantiate` is
panicking, which I think is the right course of action as opposed to
assigning arbitrary values.
* The access to the `MapView` is done via `get_mut_or_default`, which is
better than `get`. This is not always possible since in the case of
balance 0 we need to erase the key.
* The `execute_operation` is returning a non-trivial response. This is a
little bit at odds with the use of `GraphQL` API for accessing the
storage.

## Test Plan

The CI. A test has been added.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.