Correct the "tests and docs" test in the documentation. (#4250)
## Motivation

The merging of the PR #4210 led to some problems because the `Operation`
and `Abi` types are no longer
accessible when publishing. The use of example types is fine only for
tests, not for benchmarks.

## Proposal

Put back the `NativeFungibleTokenAbi` and `FungibleTokenAbi` in the
`linera-sdk::abis::fungible` and rename as needed.

## Test Plan

The `script/test_publish.sh` was used.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.