Allow contract to create blobs (#4247)
## Motivation

In order for the EVM to be able to create contracts, it needs to be able
to create the corresponding data blobs.
However, for publishing application, we need blobs that are not data
blobs and so the `publish_module` is also added.

Fixes #3707 

## Proposal

The following was done:
* Add the `create_data_blob` functionality to the runtime, thereby
addressing the issue.
* Add the `publish_bytecode` functionality to the runtime so that smart
contracts can create modules.
* The use of `fn compress` on the web forced the insertion of a relevant
function using `ruzstd` and the corresponding version bump.

## Test Plan

The CI, and 4 tests that have been introduced.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None