Remove associated types Keys and KeyValues in KV store traits (#4246)
## Motivation

* Simplify code.

As pointed out by teammates, this optimization is only used for
`DynamoDb` and doesn't apply to the databases that we use in production.

## Proposal

* Remove associated types Keys and KeyValues
* Remove the traits

I stop short of removing the iterator objects in DynamoDb. We could
probably simplify the code some more but it's non-trivial and could be
slightly slower.

## Test Plan

CI