Add event skip lists (#4263)
## Motivation

We want to be able to leave gaps in chains publishing events, just like
we do with sender chains.

## Proposal

We need skip lists for events in blocks, analogous to the ones we have
for messages: for each emitted event, have the hash and height of the
last block that emitted an event to the same stream.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)