Provide access to further linera functionalities in the EVM. (#4213)
## Motivation

We want the EVM applications to be able to fully access the Linera
functionalities.
Here we extend it by adding access to the Linera balances, authenticated
signer
as well as other features.

## Proposal

The work simply adds the following functionalities to EVM applications:
* chain_timestamp
* chain_balance
* owner_balance
* owner_balances
* balance_owners
* authenticated_signer
* authenticated_caller_id
* query_service

As it was it looked simple but several problems showed up:
* New types have to be added, but since we do not have a type `uint256`
in YAML file, we have to use a `[u8; 32]`.
* This forces a conversion to `uint256` to do in the code. The problem
is that this externalize some types, so we have an `AccountOwnerBalance`
and `AccountOwnerBalanceInner`.
* In order to force having all the types in `LineraTypes` and so in
order to have a `LineraTypes.AccountOwnerBalance` and not a
`Linera.AccountOwnerBalance`, we patch the generated file. This is added
to the CI.
* The serialization of the U256 is incorrect and returns 33 bytes
instead of 32. In order to deal with that, a type `AmountU256` is
introduced to encapsulate them.

## Test Plan

The CI and the additional tests.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.