Add basic BlockProcessor metrics. (#4244)
## Motivation

We want to see whether BlockProcessor is running (and how well).

## Proposal

Add the following metrics:
- `get_blob_histogram` 
- `get_certificate_histogram`
- `get_canonical_block_histogram`
- `block_processor_state_save_histogram` - time it takes to persiste
block processor's state to storage
- `dispatch_block_histogram` - time between we send a block to a
validator and we get an OK answer
- `dispatch_blob_histogram` - time between we send a blob to a validator
and we get an OK answer
- `destination_state_counter` - an index of (successfully) exported
block (as seen by the canonical chain) the a specific destination
- `validator_exporter_queue_length` - number of new blocks in the
(unbounded) channel between processor and the destination or otherwise
the length of a backlog of blocks we need to send to the destination

## Test Plan

Manual.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- These changes should be backported to the latest `devnet` branch
- These changes should be backported to the latest `testnet` branch

## Links


- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)