CI: fix test-crates-and-docrs job (#4281)
- **CI: fix ci job by adding linera-metrics**
- **CI: small script fix to use better interpreter**

## Motivation

As explained in the PR
(https://github.com/linera-io/linera-protocol/issues/4280) the
`test-crates-and-docrs` is now broken. This will fix it.

There's also a small nit change about using a better shebang for the
script that is used in the test.

## Proposal

Adds the `linera-metrics` package to the list of locally pushed packages
in the local registry.

## Test Plan

By adding this to GH we'll verify that it's working as intended in this
PR testing phase.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.
- These changes should be backported to the latest `devnet` branch, then
    - be released in a new SDK,
    - be released in a validator hotfix.
- These changes should be backported to the latest `testnet` branch,
then
    - be released in a new SDK,
    - be released in a validator hotfix.
- Backporting is not possible but we may want to deploy a new `devnet`
and release a new
      SDK soon.

## Links

https://github.com/linera-io/linera-protocol/issues/4280

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)