Move metrics to ClientContext (#4264)
## Motivation

Until we have a Control Pane cluster that has a Prometheus Server that
we can use to export client metrics to, we need another way to track
timings on the client

## Proposal

This PR moves the functionality of the existing timings specific to
`benchmark` into `ClientContext`, so we can keep track of more stuff
moving forward. It also makes the timings optional. We can add more to
this, then make them into Prometheus metrics in the future.

## Test Plan

Ran the client locally, saw the timings printed

## Release Plan

- Nothing to do / These changes follow the usual release cycle.