Provide the access to `created_blobs` for the contract calls. (#4278)
## Motivation

In the EVM framework of doing operation, we need to be able to do the
operations
```rust
let module_id = runtime.publish_module(....
let application_id = runtime.create_application(module_id, ....)
runtime.call_application(application_id, ....)
```

in the same block. This is currently not possible, and this PR addresses
that problem from the protocol viewpoint. A Wasm test is provided, and
the EVM should follow thereafter.

## Proposal

The following changes are made:
* When the `create_application` is made, the blobs are checked for
existence. Given that those blobs currently exist only in
`created_blobs`, we have to provide access to them.
* When the `runtime.call_application` is being made, the corresponding
contract may exist only in `created_blobs`, so again we provide access
to it.

We also provide this for the service, since a contract can call its
service during its instantiation (one example is `ethereum-tracker`).
However, the service will see the state before the instantiation, not in
the block that is currently being constructed.

## Test Plan

The CI.

A new test, `create-and-call`, has been introduced that demonstrates
those capabilities.


## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.