Add information about published events to the NewBlock notification (#4284)
## Motivation

We want to include the information about published events in the
notifications so that clients can choose whether they are interested in
the particular blocks or not.

## Proposal

Add an `event_streams` field to the `NewBlock` notification, containing
the stream IDs of streams to which the block published any events.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)
- #4261 (this is just the validator-side part, the client-side part of
actually using the event information in the notifications will be done
in a follow-up PR)