Don't panic in `send_chain_information` if the block does not exist. (#4295)
## Motivation

In #4181 `block_hashes` was changed, so it can now return a list that is
shorter than the range.

## Proposal

Don't panic if it does!

## Test Plan

@deuszx observed this panic, but I don't know how to reproduce it yet. I
added a `warn!` log, and we should investigate further if it happens
again, or find out whether it is expected that `send_chain_information`
is called with a height that we don't actually have.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)