Use full client flow on benchmarks (#4265)
## Motivation

We're changing things to use the full client flow from benchmarks. This
is part of a whole refactor to make the benchmarks more "realistic".

## Proposal

By using the full flow, we also had to rewrite how we do the client
logging of timing metrics. We'll have a timing tokio task running that
will manage the histograms, and whoever wants to log the time of
something sends a message to this tokio task.
Then every some number of seconds we print that information to the
screen for the user.

## Test Plan

Ran locally, saw the metrics being correctly printed

## Release Plan

- Nothing to do / These changes follow the usual release cycle.