Dockerize setup (#4291)
## Motivation

Currently, testing that all of the components work correctly is a
difficult and complicated task.

## Proposal

Add a Dockerfile that builds the pieces necessary to run all of the
components locally:
- validator network
- storage service
- block exporter
- indexer destination

and allows for running it with docker compose.

## Test Plan

Manual. To test, follow the new README instructions.

## Release Plan

- Nothing to do 

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)