`linera-core`: make sure to use Web-compatible time primitives (#4302)
## Motivation

https://github.com/linera-io/linera-protocol/pull/4265 moved some uses
of `std::time` primitives out of benchmarking-specific code and into the
main codepath of `linera-core`. Unfortunately, `std::time` primitives
panic on Wasm, so we must always remember to use `linera_base::time` for
anything that may be compiled for the Web (or, as a rule of thumb,
anywhere in the codebase: I don't think there's any harm to it).

## Proposal

Make sure to use the Web-compatible time shim in `linera-base`.

## Test Plan

Test in Web demos.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

Thankfully we haven't released a devnet with this bug yet.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)