Display more details. (#4305)
## Motivation

Currently only a few fields (from the header, body or messages) are
stored in a denormalized way in a database – ie they are stored in
dedicated fields – while the whole block is stored as a blob data. We
want to show more details to user, this could be done in one of the two
ways:
1. load the blob, deserialize in frontend
2. store in denormalized way and load directly in frontend

## Proposal

Implement the approach `#2` – add more fields to database tables and
store block's details directly there. Also, use the new data to show on
the frontend of the explorer.


Some refactoring was done as part of the PR:
- add a dedicated type `ExpandableSection` that shows only the metadata
of the section and expands on click to show more
- add `CopyableHash` to present a text that can be copied.

## Test Plan

Manual

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)