Make chains transfer to each other across benchmark processes (#4269)
## Motivation

Right now the benchmark client processes only do transfer amongst chains
in their wallets. We should have a way of these processes be able to
send transfers to chains in other wallets.

## Proposal

Pass a file to the benchmark client processes with all the chain IDs
(from across wallets) that we can send transfers to. Enable that via
`multi-benchmark` with an option

## Test Plan

Ran this with the new option locally against a local network, seems to
do what it's supposed to.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.