Properly log metrics for confirmed certificates (#4323)
## Motivation

Right now there's a few things wrong with how we're logging metrics for
confirmed certificates:
1. Even if processing the confirmed block fails, we still log the
transactions and everything from the block, as if it succeeded.
2. If the certificate has already been processed, we log the
transactions and everything else again, as if it's a new block, which
causes over reporting.

## Proposal

Fix the two issues with the logging.

## Test Plan

CI + will test in a deployed network that I don't see the over reporting
anymore.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.