Make `ContractRuntime::transfer` synchronous for local transfers. (#4268)
## Motivation

During the implementation of `transfer` operations in the EVM, we face
the problem that while the
`ContractRuntime::transfer` immediately debits the source, and the
destination receives the amount
only on the next block.

While this makes complete sense if the destination chain is different
from the source chain, there is a use
case for transfer operations within the same chain that are done
immediately without emitting messages.

Fixes #3486 

## Proposal

The code of `SystemExecutionStateView::transfer` has been changed so
that it does not emit a message if the source chain is the same as the
destination chain.

The operations that have been corrected are:
* `Claim`
* `Transfer`
* `Withdraw`

## Test Plan

The CI.

Several tests in `worker_test.rs` had to be changed because the emission
of messages had changed.

The end-to-end tests are also unified because now `fungible` and the
native token have the same behavior for the emission of messages.

More tests for `Claim`, `Withdraw`, and `Credit` would be useful to do
in subsequent PRs.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.