One central benchmark command (#4338)
## Motivation

Right now we have `linera benchmark` and `linera multi-benchmark`. It
would be nice if all the benchmark stuff was centralized in one command,
with sub commands.

## Proposal

Modify things so that we have one single `linera benchmark` command, but
with subcommands:
* `linera benchmark single`
* `linera benchmark multi`

## Test Plan

CI should cover it, but will test it against a local network as well

## Release Plan

- Nothing to do / These changes follow the usual release cycle.