Eliminate the `Recipient` type. (#4343)
## Motivation

The Recipient type variant `Burn` is used only for tests. So, we remove
it. This let the `Recipient` having just a variant `Account`. And so the
`Recipient` type is eliminated.

Fixes #3580 

## Proposal

Straightforward from the description.

## Test Plan

The CI.
The tests were restructured accordingly. For some tests the balance is
now unchanged. For some tests we introduced another chain.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None.