Merge "attempted" and "temporary" changes into a single `ChainWorkerGuard` type. (#4354)
## Motivation

The distinction between `ChainWorkerStateWithAttemptedChanges` and
`ChainWorkerStateWithTemporaryChanges` is unnecessary: The `succeeded`
field might as well be omitted, because calling `rollback` after saving
is a no-op. It's also a bit misleading, because if one calls `save` in
`ChainWorkerStateWithTemporaryChanges`, the changes _will_ be permanent.

## Proposal

Remove `succeeded`. Replace the two types with a single
`ChainWorkerGuard`.

This will enable other cleanups, but for clarity I'd do those in a
separate PR. In this one, the methods have just been moved together into
a single type.

## Test Plan

CI

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)