Don't log errors returned from gRPC client methods. (#4365)
## Motivation

An error in the gRPC client can usually be caused by just one faulty
validator, which is not considered to be an error for the overall BFT
system.

## Proposal

Reduce the log level. (Thanks for the proper fix, @ndr-ds!)

## Test Plan

CI

And we will need to see in the logs if there are any other unwanted
error-level logs.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

- Closes #3391.
- [reviewer
checklist](https://github.com/linera-io/linera-protocol/blob/main/CONTRIBUTING.md#reviewer-checklist)

---------

Co-authored-by: Andre da Silva <andre.dasilva@linera.io>