Make the `clone_unchecked` a function that does not return errors. (#4362)
## Motivation

An examination of the source code reveals that `clone_unchecked` never
returns an error.
The case of `?` are due to the calls to `clone_unchecked`.

## Proposal

Remove the error case.

Note that there are second order effects. For example `fn
chain_state_view` no longer needs to return an error.
But this is for subsequent works.

## Test Plan

CI,.

## Release Plan

- Nothing to do / These changes follow the usual release cycle.

## Links

None